18: accessing current user
Instead of using a helper, this uses ActiveSupport::Current.

I really love this! It seems like "money laundering for global state" in the best way possible.

For each request, we get to build up global state for that request. And we know it will be clean the next request.

I think this is perfect for centralizing data that would be used for permission / ability / policy checks on the user.